Year 2019 Day 12

+++ src/main/java/net/eugenpaul/adventofcode/y2018/day18/Day18.java
@@ -71,7 +71,7 @@ public class Day18 extends SolutionTemplate {
 
             String hash = toHash(area);
 
-            if (storage.add(hash, i)) {
+            if (storage.add(hash, (long) i)) {
                 break;
             }
         }
